% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AcMtEst.R
\name{AcMtEst}
\alias{AcMtEst}
\title{Combine Acoustic and Midwater Trawl Survey Data}
\usage{
AcMtEst(SimPop, AcMtSurv, AcExcl = c(0, 0), MtExcl = c(0, 0),
  PanelProps = c(0.4, 0.3, 0.2, 0.1), SelecParam = NULL, Seed = NULL)
}
\arguments{
\item{SimPop}{A list with elements \code{LakeInfo}, \code{FishInfo}, \code{FishParam},
  \code{FishPop}, typically output from \code{\link{SimFish}}.
See \code{\link{SimFish}} for details on the list elements.}

\item{AcMtSurv}{A list with elements \code{Targets}, \code{AcSummaryCell},
\code{AcSummaryColumn}, \code{MtCatch},
typically output from \code{\link{SampFish}}.}

\item{AcExcl}{A numeric vector of length 2, depth of acoustic "dead" zones at the surface
and at the bottom (in m), default of c(0, 0) represents 100\% acoustic
availability of fish.}

\item{MtExcl}{A numeric vector of length 2, depth of zones unfishable with the midwater
trawl at the surface and at the bottom (in m), default of c(0, 0)
represents 100\% midwater trawl availability of fish.}

\item{PanelProps}{A numeric vector of length 4, size of the different mesh panel zones of the
  midwater trawl, mouth (outermost), middle, aft, and cod (inner),
  default c(0.4, 0.3, 0.2, 0.1).
Sizes are expressed as proportions of the distance from the outer edge of
  the trawl to the trawl center in both the vertical and horizontal
  directions, and they should add up to 1.  Use \code{\link{ViewZones}}
  to visualize the mesh panel zones.}

\item{SelecParam}{A data frame with 6 columns in which each row provides the midwater trawl
  selectivity parameters for a given fish group and mesh panel zone.
All columns must be completely filled in (no missing values).
Selectivity is assumed to be 100\% for any group-zone combination not
  represented as a row in the data frame.
For 100\% selectivity of small fish, use \code{MtL50Small = -Inf} and
  any slope.
For 100\% selectivity of large fish, use \code{MtL50Large = Inf} and
  any slope.
Column names and descriptions:
\itemize{
  \item \code{G} = character, a one-letter nickname for the group
    (e.g., fish species and life stage) used in plotting
  \item \code{Zone} = character, mesh panel zone, one of "mouth", "middle",
    "aft", or "cod"
  \item \code{MtL50Small} = numeric, the length (in mm) at which small fish
    have a 50\% probability of being captured by the trawl
  \item \code{MtSlopeSmall} = numeric, the (inverse) slope at which small
    fish probability of capture increases with length, smaller values are
    steeper
  \item \code{MtL50Large} = numeric, the length (in mm) at which large fish
    have a 50\% probability of being captured by the trawl
  \item \code{MtSlopeLarge} = numeric, the (absolute value of the inverse)
    slope at which large fish probability of capture decreases with length,
    smaller values are steeper
}}

\item{Seed}{An integer scalar, starting seed for stochasticity incorporated in acoustic
  and midwater trawl catchability.
Use \code{Seed} to ensure the same individual fish are included in the
  surveys with each call to \code{CatchComb}.
Otherwise, if set to NULL, the default, a random seed is used, resulting
  in a different fish selection with each call to \code{CatchComb}.}
}
\value{
A data frame with estimated fish density (in number per ha) and biomass
    (in kg per ha) for each sampling event and group (species, lifestage).
}
\description{
Combine survey data from acoustic transects and midwater trawl tows (created
  by \code{\link{SampFish}}).
Apply availability to the acoustic data and catchability (availability and
  selectivity) to the midwater trawl catch.
}
\details{
A classification tree is used to relate the catch composition of
    the midwater trawl to the location of the trawl in the lake
    (e.g., MTReast, ACnorth, MTRd2sh, MTRbdep).  This tree is then used
    to assign a single midwater trawl catch to each acoustic cell
    (interval x layer), such that the estimated acoustic densities can be
    assigned to specific fish groups (species, life stages).
    See, for example, Yule et al. (2013).
}
\examples{
# parameters for small (a) and large (A) alewife as input to the simulator
fishp <- data.frame(
  G = c("a", "A", "A"),
	 Z = c(50, 140, 140), ZE = c(0.25, 0.2, 0.2),
	 LWC1 = 0.000014, LWC2 = 2.8638, LWCE = 0.18,
	 TSC1 = -64.2, TSC2 = 20.5, TSCE = c(0.02, 0.07, 0.07),
	 PropN = c(0.55, 0.25, 0.20),
	 E = c(NA, 900, 2800), EE = c(NA, 4.5, 0.3),
	 N = NA, NE = NA,
	 WD = c(5, 15, 15), WDE = c(0.5, 0.7, 0.7),
	 D2B = NA, D2BE = NA
)

# simulate the fish population
res <- SimFish(LakeName="Clear Lake", LkWidth=3000, LkLength=2000,
	 BotDepMin=20, BotDepMax=100, FishParam=fishp, TotNFish=50000)

# survey the population
surv <- SampFish(SimPop=res, NumEvents=2, AcNum=5, AcInterval=3000,
	 AcLayer=10, AcAngle=7, MtNum=25, MtHt=10, MtWd=10, MtLen=200)

selec <- data.frame(
	 G = c("A", "a", "A", "a", "A", "a"),
	 Zone = c("mouth", "mouth", "middle", "middle", "aft", "aft"),
	 MtL50Small = c(100, 100, 60, 60, 30, 30),
	 MtSlopeSmall = c(40, 40, 30, 30, 20, 20),
	 MtL50Large = c(180, 180, Inf, Inf, Inf, Inf),
	 MtSlopeLarge = c(20, 20, 100, 100, 100, 100)
)

AcMtEst(SimPop=res, AcMtSurv=surv, Seed=927)
AcMtEst(SimPop=res, AcMtSurv=surv, AcExcl=c(5, 10),
	 MtExcl=c(2, 2), SelecParam=selec, Seed=204)
}
\references{
Yule, DL, JV Adams, DM Warner, TR Hrabik, PM Kocovsky, BC Weidel, LG Rudstam,
  and PJ Sullivan.  2013.
  Evaluating analytical approaches for estimating pelagic fish biomass using
    simulated fish communities.
  Canadian Journal of Fisheries and Aquatic Sciences 70:1845-1857.
  \emph{http://www.nrcresearchpress.com/doi/abs/10.1139/cjfas-2013-0072#.U1KYxPldXTQ}
}
\seealso{
\code{\link{SimFish}}, \code{\link{SampFish}}, \code{\link{ViewZones}},
    \code{\link{TuneSelec}}.
}

