% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AcSmry}
\alias{AcSmry}
\title{Summarize Acoustic Survey Data}
\usage{
AcSmry(AcTarg, LakeInfo, SurvParam)
}
\arguments{
\item{AcTarg}{A data frame list with information on fish detected in an acoustics survey, specifically the \code{Targets} data frame
output from \code{\link{SampFish}}.  Each row represents a single target, columns describe the specific location of fish in the lake and their target strengths.}

\item{LakeInfo}{A list with the lake inputs supplied as arguments to \code{\link{SimFish}} as well as a few additional objects.}

\item{SurvParam}{A named vector with the survey inputs supplied as arguments to \code{\link{SampFish}}.}
}
\value{
A list with 2 elements.
\itemize{
  \item \code{AcCell} is a data frame with information on the acoustic survey summarized by acoustic interval and layer.
  \item \code{AcColumn} is a data frame with information on the acoustic survey summarized by acoustic interval.
}
}
\description{
Summarize acoustic survey data by interval and layer.
}
\details{
Acoustic intervals are identified by the easting (in m) of their midpoint.
Acoustic layers are indentified by the water depth (in m) of their midpoint.

A weighting variable, range weight (Yule 2000), is used to account for different volumes of water
sampled in the acoustic survey as a function of the distance from the transducer (in m) and the transducer half angle (0.5 * \code{AcAngle}).
The sum of the range weights is reported as \code{sum.rw} in the \code{AcCell} and \code{AcColumn} data frames.
}
\examples{
# parameters for small (a) and large (A) alewife as input to the simulator
fishp <- data.frame(
	G = c("a", "A", "A"),
	Z = c(50, 140, 140), ZE = c(0.25, 0.2, 0.2),
	LWC1 = 0.000014, LWC2 = 2.8638, LWCE = 0.18,
	TSC1 = -64.2, TSC2 = 20.5, TSCE = c(0.02, 0.07, 0.07),
	PropN = c(0.55, 0.25, 0.20),
	E = c(NA, 900, 2800), EE = c(NA, 4.5, 0.3),
	N = NA, NE = NA,
	WD = c(5, 15, 15), WDE = c(0.5, 0.7, 0.7),
	D2B = NA, D2BE = NA)

# simulate the fish population
res <- SimFish(LakeName="Clear Lake", LkWidth=3000, LkLength=2000, BotDepMin=20, BotDepMax=100,
	FishParam=fishp, TotNFish=50000, Seed=667)

# survey the population
surv <- SampFish(SimPop=res, NumEvents=2, AcNum=5, AcInterval=3000, AcLayer=10, AcAngle=7, MtNum=25, MtHt=10, MtWd=10, MtLen=200, Seed=545)

AcSmry(AcTarg=surv$Targets, LakeInfo=res$LakeInfo, SurvParam=surv$SurvParam)
}
\references{
Yule, DL.  2000.
\href{http://www.tandfonline.com/doi/abs/10.1577/1548-8675(2000)020%3C0759%3ACOHAAP%3E2.3.CO%3B2}{Comparison
of horizontal acoustic and purse-seine estimates of salmonid densities and sizes in eleven Wyoming waters}.
North American Journal of Fisheries Management 20:759-775.
}
\seealso{
\code{\link{SampFish}}
}

