% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfromx}
\alias{dfromx}
\title{Determine Distance to Shore from Easting}
\usage{
dfromx(x, d2shr.we, eastr)
}
\arguments{
\item{x}{A numeric vector of the eastings to convert (in m).}

\item{d2shr.we}{A numeric vector of length 2, distance (in the "x" direction) from west and east shores excluded from lake (in m).}

\item{eastr}{A numeric vector of length 2, easting range (minimum and maximum) corresponding to the west and east shores of the lake (in m).}
}
\value{
A numeric vector, same length as \code{x}, with distances to shore (west or east, whichever is closer) for each easting provided (in m).
}
\description{
Determine the distance to the lake shore from a given easting.
}
\details{
An internal function called by \code{\link{SimFish}()}.
}
\seealso{
\code{\link{SimFish}}, \code{\link{zfromx}()}, \code{\link{xfromz}()}.
}

