% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SimFish}
\alias{SimFish}
\title{Simulate a Fish Population}
\usage{
SimFish(LakeName, LakeNick, LkWidth = 30000, LkLength = 20000,
  BotDepMin = 20, BotDepMax, BotDepVertex = 2 * BotDepMax, FishParam,
  TotNFish, TSRange = c(-65, -20), plots.pdf = NA, Seed = NULL)
}
\arguments{
\item{LakeName}{A character scalar, full name for artificial lake to be used in plot titles.}

\item{LakeNick}{A character scalar, nickname for the artificial lake to be used in file names.}

\item{LkWidth}{A numeric scalar, the width of the lake in the west-east direction (in m), default 30,000.}

\item{LkLength}{A numeric scalar, the length of the lake in the south-north direction (in m), default 20,000.}

\item{BotDepMin}{A numeric scalar, the minimum bottom depth of the lake, at either the west or east shoreline (in m), default 20.}

\item{BotDepMax}{A numeric scalar, the maximum bottom depth of the lake (in m).}

\item{BotDepVertex}{A numeric scalar, the vertical distance from the surface to the "vertex" of the lake bottom (in m), default \code{2*BotDepMax}.
The "vertex" of the lake bottom is the point at which the angled lake beds along the west and east shores would intersect,
were they not cut off first by the specified \code{BotDepMax}.}

\item{FishParam}{A data frame with 19 columns in which each row describes a sub-population of fish to be placed in the artificial lake.
The first 11 columns must be completely filled in (no missing values).
The last 8 columns may have some missing values.
However, in each row, either water depth (WD and WDE) or distance to bottom (D2B and D2BE) \strong{must} be filled in, but \strong{not both}.
Column names and descriptions:
\itemize{
  \item \code{G} = character, a one-letter nickname for the group (e.g., fish species and lifestage) that will be used in plotting
  \item \code{Group} = character, full name of the group (e.g., fish species and lifestage)
  \item \code{Z} = numeric, mean length (in mm)
  \item \code{ZE} = numeric, error around mean length, expressed as SD/mean
  \item \code{LWC1}, \code{LWC2} = numeric, length-weight regression coefficients, where wt = LWC1*len^LWC2, (wt in g, len in mm)
  \item \code{LWCE} = numeric, error around weight estimate, expressed as SD(estimate)/estimate
  \item \code{TSC1}, \code{TSC2} = numeric, target strength and length relation coefficients, ts = TSC1 + TSC2*log10(len/10), (ts in db, len in mm)
  \item \code{TSCE} = numeric, error around target strength estimate, expressed as SD(estimate)/estimate
  \item \code{PropN} = numeric, approximate proportion of population that the row represents (automatically adjusted to ensure they sum to 1)
  \item \code{E} = numeric, mean easting (km)
  \item \code{EE} = numeric, error around easting, expressed as SD/mean
  \item \code{N} = numeric, mean northing (m)
  \item \code{NE} = numeric, error around northing, expressed as SD/mean
  \item \code{WD} = numeric, mean water depth (m)
  \item \code{WDE} = numeric, error around water depth, expressed as SD/mean
  \item \code{D2B} = numeric, mean distance to bottom (m)
  \item \code{D2BE} = numeric, error around distance to bottom, expressed as SD/mean
}}

\item{TotNFish}{A numeric scalar indicating the target number of fish to put in the lake.
The actual number of fish in the population will likely be smaller than \code{TotNFish}, because the process used to populate the lake with fish
ends up with some fish out of water (beyond the boundaries of the artificial lake), which are then removed from the population.
Memory on your computer will limit the size of \code{TotNFish}.}

\item{TSRange}{A numeric vector of length 2, the range of target strengths to use for the fish (in db), default c(-65, -20).}

\item{plots.pdf}{A character scalar, name of pdf file to store the diagnostic plots in.  If NA, the default,
plots will be displayed on the screen instead.  If FALSE, no plots will be created.}

\item{Seed}{An integer scalar, starting seed for stochasticity incorporated in fish population generation.
Provide an integer to ensure the same population is generated with each call to \code{\link{SimFish}()}.
Otherwise, if set to NULL, the default, a random seed will be used, resulting in a different population with each call to \code{\link{SimFish}()}.}
}
\value{
A list with five elements, (1) \code{Truth}, a data frame (or matrix?) with the total number and weight
of each fish group in the population,
(2) \code{LakeInfo}, a list with the lake inputs supplied as arguments to \code{SimFish} as well as a few additional objects
which will be useful in surveying the population,
\itemize{
  \item \code{ints} = a numeric vector of length 2, the intercepts of the angled lake beds along the west and east shores of the lake (in m)
  \item \code{slopes} = a numeric vector of length 2, the slopes of the angled lake beds along the west and east shores of the lake (unitless)
  \item \code{d2shr.we} = a numeric vector of length 2, distance (in the "x" direction) from west and east shores excluded from lake (in m)
  \item \code{mid.d} = ???,
}
(3) \code{FishInfo}, a list with the fish inputs supplied as arguments to \code{SimFish},
(4) \code{FishParam}, the data frame supplied as an argument to \code{SimFish}, and
(5) \code{FishPop}, a data frame in which each row is a fish, and the 10 columns describe the fish group (\code{sp}), location
(easting \code{f.east}, northing \code{f.north}, distance to shore \code{f.d2sh}, bottom depth \code{f.botdep},
water depth \code{f.fdep}, distance to bottom \code{f.d2bot}, all in m), and
fish size (total length in mm \code{len}, weight in g \code{wt}, and target strength in db, \code{ts}).
}
\description{
Create a simulated population of pelagic fish in an artificial lake.
}
\details{
Describe the lake shape.  Refer to two figures (top and side views).
You may wish to cap the total number of fish at 5 million for a memory of ~ 2 GB (2047 MB).
This limit can be increased if you have more memory available in R.
You can check the memory available with \code{\link{memory.limit}()}.
}

