% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TuneSelec}
\alias{TuneSelec}
\title{Tune Midwater Trawl Selectivity}
\usage{
TuneSelec()
}
\description{
Interactive visualization of midwater trawl selectivity curve for aid in tuning selectivity parameters used in \code{\link{AcMtEst}}.
}
\details{
Interactive sliders are provided for the four parameters of interest which are required for defining the midwater trawl selectivity
of the mesh panel zones as a double logistic curve:
\itemize{
  \item \code{MtL50Small} the length (in mm) at which small fish have a 50\% probability of being captured by the trawl
  \item \code{MtSlopeSmall} the slope at which small fish probability of capture increases with length
  \item \code{MtL50Large} the length (in mm) at which large fish have a 50\% probability of being captured by the trawl
  \item \code{MtSlopeLarge} the (absolute value of the) slope at which large fish probability of capture decreases with length
	}

Note that the sliders pop up in a separate R window, which may be hidden if you click on another window.
}
\examples{
\dontrun{
TuneSelec()
}
}
\seealso{
\code{\link{AcMtEst}}, \code{\link{logit2}}
}

