% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfromz.R
\name{xfromz}
\alias{xfromz}
\title{Determine Easting from Bottom Depth}
\usage{
xfromz(z, maxz, ints, slopes, shore = "random")
}
\arguments{
\item{z}{A numeric vector of the distances to lake bottom to convert (in m).}

\item{maxz}{A numeric scalar, the maximum bottom depth of the lake (in m).}

\item{ints}{A numeric vector of length 2, the intercepts of the angled lake beds along
the west and east shores of the lake (in m).}

\item{slopes}{A numeric vector of length 2, the slopes of the angled lake beds along
the west and east shores of the lake (unitless).}

\item{shore}{A numeric or character scalar, indicating if the shore should be west (0),
east (1) or randomly assigned ("random", default).}
}
\value{
A numeric vector, same length as \code{x}, with distances to bottom for
  each easting provided (in m).
}
\description{
Determine the easting from a given distance to lake bottom,
incorporating random assignment to west or east shore if not specified.
}
\details{
An internal function called by \code{\link{SimFish}}.
}
\seealso{
\code{\link{SimFish}}, \code{\link{dfromx}}, \code{\link{zfromx}}.
  The assignment to shore is necessary because the artificial lake will have
  two eastings for each bottom depth (except for the depth at the vertex).
}

