% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zfromx}
\alias{zfromx}
\title{Determine Bottom Depth from Easting}
\usage{
zfromx(x, maxz, eastr, ints, slopes)
}
\arguments{
\item{x}{A numeric vector of the eastings to convert (in m).}

\item{maxz}{A numeric scalar, the maximum bottom depth of the lake (in m).}

\item{eastr}{A numeric vector of length 2, easting range (minimum and maximum) corresponding to the west and east shores of the lake (in m).}

\item{ints}{A numeric vector of length 2, the intercepts of the angled lake beds along the west and east shores of the lake (in m).}

\item{slopes}{A numeric vector of length 2, the slopes of the angled lake beds along the west and east shores of the lake (unitless).}
}
\value{
A numeric vector, same length as \code{x}, with distances to bottom for each easting provided (in m).
}
\description{
Determine the distance to the bottom from a given easting.
}
\details{
An internal function called by \code{\link{SimFish}}.
}
\seealso{
\code{\link{SimFish}}, \code{\link{dfromx}}, \code{\link{xfromz}}.
}

